# RFC: Specialize cdf for discrete distributions with finite support