* For discrete functions with finite support, calculate
  `cdf` from right if we're in the right half of that support
* Remove cdf specialized on `BetaBinomial` (handled by
  `DiscreteUnivariateDistribution` version)
* ~~Remove extraneous definitions of `*cdf` functions which simply
  convert values to `Int` or `Float64`, where these conversions will
  happen in a call to `cdf(...)`~~ See #428