@kmsquire 2016-01-10
I took out the removal of extraneous definitions.  The `Float64` versions of those were handled by #428 more completely, and the `Int` version can be submit separately.