@kmsquire 2016-01-10
Better title.  This now directly handles the TODO in #422.