@kmsquire 2016-01-10
The push failure looks like a timeout, and everything passed in the PR tests.

@andreasnoack--this implements a more general version of your suggestion [here](https://github.com/JuliaStats/Distributions.jl/pull/422#discussion_r49270240).

The changes here should not have a negative impact on the performance of calculating the cdf (and should help for large `n`). I'm going to do a few quick timing tests today to make sure, and then merge if there aren't any addtional comments.