@kmsquire 2016-01-10
Seem to be no negative consequences on timing for this PR, and the `cdf` of the upper half of discrete distributions with finite support now is now a bit faster.

(Timings below are after warmup.  Most relevant is first line of each benchmark results.)

Master:
```julia
julia> bbn = BetaBinomial(100, 0.2, 0.25)
Distributions.BetaBinomial(n=100, α=0.2, β=0.25)

julia> cdf(bbn, 1)
0.259735928585796

julia> @benchmark cdf(bbn, 50)
================ Benchmark Results ========================
     Time per evaluation: 130.01 μs [118.37 μs, 141.66 μs]
Proportion of time in GC: 2.86% [1.46%, 4.26%]
        Memory allocated: 40.02 kb
   Number of allocations: 1364 allocations
       Number of samples: 3901
   Number of evaluations: 82201
         R² of OLS model: 0.535
 Time spent benchmarking: 10.84 s


julia> @benchmark cdf(bbn, 75)
================ Benchmark Results ========================
     Time per evaluation: 199.53 μs [178.40 μs, 220.65 μs]
Proportion of time in GC: 3.09% [1.52%, 4.65%]
        Memory allocated: 60.73 kb
   Number of allocations: 2097 allocations
       Number of samples: 3401
   Number of evaluations: 50701
         R² of OLS model: 0.486
 Time spent benchmarking: 10.37 s


julia> bbn = BetaBinomial(10000, 0.2, 0.25)
Distributions.BetaBinomial(n=10000, α=0.2, β=0.25)

julia> @benchmark cdf(bbn, 5000)
================ Benchmark Results ========================
     Time per evaluation: 1.10 s [1.08 s, 1.12 s]
Proportion of time in GC: 0.21% [0.00%, 0.84%]
        Memory allocated: 9.29 mb
   Number of allocations: 658247 allocations
       Number of samples: 8
   Number of evaluations: 8
 Time spent benchmarking: 9.94 s


julia> @benchmark cdf(bbn, 7500)
================ Benchmark Results ========================
     Time per evaluation: 2.00 s [1.94 s, 2.06 s]
Proportion of time in GC: 0.23% [0.00%, 0.91%]
        Memory allocated: 14.31 mb
   Number of allocations: 1071488 allocations
       Number of samples: 3
   Number of evaluations: 3
 Time spent benchmarking: 7.99 s


julia> nbm = NegativeBinomial(1, 0.0001)
Distributions.NegativeBinomial(r=1.0, p=0.0001)

julia> @benchmark cdf(nbm, 10000)
================ Benchmark Results ========================
     Time per evaluation: 313.12 ns [311.22 ns, 315.01 ns]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 7001
   Number of evaluations: 1580801
         R² of OLS model: 0.993
 Time spent benchmarking: 0.54 s
```

This PR:
```julia
julia> bbn = BetaBinomial(100, 0.2, 0.25)
Distributions.BetaBinomial(n=100, α=0.2, β=0.25)

julia> cdf(bbn, 1)
0.259735928585796

julia> @benchmark cdf(bbn, 50)
================ Benchmark Results ========================
     Time per evaluation: 130.38 μs [118.07 μs, 142.70 μs]
Proportion of time in GC: 2.90% [1.48%, 4.32%]
        Memory allocated: 40.07 kb
   Number of allocations: 1366 allocations
       Number of samples: 3901
   Number of evaluations: 82201
         R² of OLS model: 0.509
 Time spent benchmarking: 10.83 s


julia> @benchmark cdf(bbn, 75)
================ Benchmark Results ========================
     Time per evaluation: 52.27 μs [47.95 μs, 56.59 μs]
Proportion of time in GC: 2.61% [1.38%, 3.83%]
        Memory allocated: 18.53 kb
   Number of allocations: 603 allocations
       Number of samples: 4801
   Number of evaluations: 194401
         R² of OLS model: 0.523
 Time spent benchmarking: 10.31 s


julia> bbn = BetaBinomial(10000, 0.2, 0.25)
Distributions.BetaBinomial(n=10000, α=0.2, β=0.25)

julia> @benchmark cdf(bbn, 5000)
================ Benchmark Results ========================
     Time per evaluation: 1.10 s [1.07 s, 1.12 s]
Proportion of time in GC: 0.22% [0.00%, 0.88%]
        Memory allocated: 9.29 mb
   Number of allocations: 658250 allocations
       Number of samples: 8
   Number of evaluations: 8
 Time spent benchmarking: 9.87 s


julia> @benchmark cdf(bbn, 7500)
================ Benchmark Results ========================
     Time per evaluation: 207.78 ms [202.53 ms, 213.03 ms]
Proportion of time in GC: 0.74% [0.00%, 2.10%]
        Memory allocated: 4.27 mb
   Number of allocations: 244834 allocations
       Number of samples: 44
   Number of evaluations: 44
 Time spent benchmarking: 9.37 s

julia> nbm = NegativeBinomial(1, 0.0001)
Distributions.NegativeBinomial(r=1.0, p=0.0001)

julia> @benchmark cdf(nbm, 10000)
================ Benchmark Results ========================
     Time per evaluation: 314.75 ns [312.51 ns, 316.98 ns]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 6901
   Number of evaluations: 1437201
         R² of OLS model: 0.990
 Time spent benchmarking: 0.50 s
```