@kmsquire 2016-01-10
Merging, since tests pass and timing is improved for some cases.  If there are concerns about the implementation, they can be handled in a future PR.